requirements.txt
transformers
torch
numpy
faiss-cpu
sentence-transformers
gradio
accelerate
gradio[oauth]